% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{get_types}
\alias{get_types}
\title{Parse an R source file with roxygen2 doc comments and extract
argument types and return value}
\usage{
get_types(path, punct = ";")
}
\arguments{
\item{path}{character(1) path to source file}

\item{punct}{character(1) regular expression of tokens to be removed, defaults to ";"}
}
\description{
Parse an R source file with roxygen2 doc comments and extract
argument types and return value
}
\examples{
srcf = system.file("demos", "mvcal.R", package="getTypeinfo")
tt = get_types(srcf)
lapply(tt, packup)
}
