% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{packup}
\alias{packup}
\title{simplify result of internal parameter processing}
\usage{
packup(x)
}
\arguments{
\item{x}{output of internal .parmtypes}
}
\description{
simplify result of internal parameter processing
}
